# doc/swagger_v1.yaml
openapi: 3.0.0
info:
  title: Office Queue API
  version: 1.0.0
  description: API for the project
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
paths:
...
  /queues:
    get:
      summary: Get all queues
      responses:
        '200':
          description: A list of queues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Queue'
    post:
      summary: Create a new queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Queue'
      responses:
        '201':
          description: Queue created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
  /queues/served/{counter_id}:
    get:
      summary: Get daily served queues for a specific counter
      parameters:
        - in: path
          name: counter_id
          required: true
          schema:
            type: integer
          description: The ID of the counter
      responses:
        '200':
          description: A list of served queues for the specified counter
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Queue'
  /queues/next/{counter_id}:
    post:
      summary: Get the next ticket in queue for a specific counter
      parameters:
        - in: path
          name: counter_id
          required: true
          schema:
            type: integer
          description: The ID of the counter
      responses:
        '201':
          description: The next queue for the specified counter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
  /queues/{ticket_id}/close:
    post:
      summary: Close a specific queue by ticket ID
      parameters:
        - in: path
          name: ticket_id
          required: true
          schema:
            type: integer
          description: The ID of the ticket whose queue is to be closed
      responses:
        '200':
          description: Queue closed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Queue'
...
components:
  schemas:
    Customer:
      type: object
      properties:
        id:
          type: integer
          example: 1
        firstName:
          type: string
          example: "Mario"
        lastName:
          type: string
          example: "Rossi"
        phoneNumber:
          type: string
          example: "391234567890"
      required: [id, firstName, lastName]
    Service:
      type: object
      properties:
        id:
          type: integer
          example: 10
        name:
          type: string
          example: "Passport"
        description:
          type: string
          example: "Passport issuance service"
        averageServiceTime:
          type: integer
          example: 30
      required: [id, name, averageServiceTime]
    Ticket:
      type: object
      properties:
        id:
          type: number
          example: 1001
        service:
          $ref: '#/components/schemas/Service'
        customer:
          $ref: '#/components/schemas/Customer'
        createdAt:
          type: string
          format: date-time
      required: [id, service, customer]
    Queue:
      type: object
      properties:
        id:
          type: integer
          example: 1
        timestamp:
          type: string
          format: date-time
        ticket:
          $ref: '#/components/schemas/Ticket'
        counter:
          $ref: '#/components/schemas/Counter'
        servedAt:
          type: string
          format: date-time
          example: "2023-10-01T10:30:00Z"
        served:
          type: boolean
          example: false
        closedd_at:
          type: string
          format: date-time
          example: "2023-10-01T11:00:00Z"
        
      required: [id, ticket]
    Counter:
      type: object
      properties:
        id:
          type: number
          example: 1
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
        queues:
          type: array
          items:
            $ref: '#/components/schemas/Queue'
      required: [id, services, queues]
...