# src/models/dto/Service.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Office Queue API
 * API for the project
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @interface Service
 */
export interface Service {
    /**
     * 
     * @type {number}
     * @memberof Service
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Service
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Service
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof Service
     */
    averageServiceTime: number;
}

/**
 * Check if a given object implements the Service interface.
 */
export function instanceOfService(value: object): value is Service {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('averageServiceTime' in value) || value['averageServiceTime'] === undefined) return false;
    return true;
}

export function ServiceFromJSON(json: any): Service {
    return ServiceFromJSONTyped(json, false);
}

export function ServiceFromJSONTyped(json: any, ignoreDiscriminator: boolean): Service {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'averageServiceTime': json['averageServiceTime'],
    };
}

export function ServiceToJSON(json: any): Service {
    return ServiceToJSONTyped(json, false);
}

export function ServiceToJSONTyped(value?: Service | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'averageServiceTime': value['averageServiceTime'],
    };
}